Updates to fix working directory management on Mac OSX when running from Finder, to remove warnings from all platforms, updates to GNUMakefile to detect installed FLTK libraries and remove dependency on FLTKDIR environment variable, updated release.txt.

+++ src/inet_pton.c
@@ -165,7 +165,7 @@ int     inet_pton4(char *src, u_char *dst, int pton)
          * Concoct the address according to
          * the number of parts specified.
          */
-        n = pp - parts + 1;
+        n = (u_int) (pp - parts + 1);
         /* inet_pton() takes dotted-quad only.  it does not take shorthand. */
         if (pton && n != 4)
                 return (0);
